vertex-search-advanced/requirements.txt
Flask==2.3.3
requests==2.31.0
debugpy # Required for debugging.
google-auth==2.3.2
google-cloud-discoveryengine==0.11.1